+++ test/methodOverride.js
@@ -33,11 +33,28 @@ describe('connect.methodOverride()', function(){
     .expect('DELETE', done);
   })
 
+  it('should work on other methods', function(done){
+    app.request()
+    .put('/')
+    .set('Content-Type', 'application/x-www-form-urlencoded')
+    .write('_method=delete')
+    .expect('DELETE', done);
+  })
+
   it('should ignore invalid methods', function(done){
     app.request()
     .post('/')
     .set('Content-Type', 'application/x-www-form-urlencoded')
     .write('_method=<whatever>')
     .expect('POST', done);
+  })
+
+  it('should support X-HTTP-Method-Override header', function(done){
+    app.request()
+    .post('/')
+    .set('Content-Type', 'application/x-www-form-urlencoded')
+    .set('X-HTTP-Method-Override', 'PUT')
+    .write('_method=DELETE')
+    .expect('PUT', done);
   })
 })